[VERY WIP] refactor: Seperate UI into sep modules
Ignore, it doesn't work
