+++ imports.py
@@ -1,0 +1,15 @@
+from PyQt6.QtWidgets import QMainWindow, QApplication, QLabel, QPushButton, QVBoxLayout, QWidget, QCheckBox, QStackedLayout, QHBoxLayout, QGroupBox, QComboBox, QProgressBar, QLineEdit, QMessageBox, QFileDialog, QInputDialog, QTreeWidget, QTreeWidgetItem
+from PyQt6.QtGui import QPixmap, QIcon, QImage
+from PyQt6.QtCore import Qt, QThread, pyqtSignal, QByteArray, QFile, pyqtSlot
+import requests
+import os
+import subprocess
+import sys
+import json
+from zipfile import ZipFile
+import shutil
+import tempfile
+from icons import styledark_rc
+import win32com.client
+import winreg
+from pathlib import Path
